Added url bat and js file, moved files to src
