+++ .gitignore
@@ -1,0 +1,8 @@
+# Config Files
+
+botsettings.json
+
+# Node modules
+
+/node_modules
+
