+++ package.json
@@ -1,0 +1,16 @@
+{
+  "name": "bissle-master",
+  "version": "1.0.0",
+  "description": "Remnant server bot",
+  "private": true,
+  "main": "index.js",
+  "scripts": {
+    "url": "node generateBotUrl.js",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "cptdrrev",
+  "dependencies": {
+    "discord.js": "^11.5.1",
+    "sqlite": "^3.0.3"
+  }
+}
