+++ src/inventory.json
