+++ start.bat
@@ -1,2 +1,2 @@
-nodemon index.js
+nodemon src/index.js
 pause
