+++ url.bat
@@ -1,0 +1,2 @@
+nodemon src/getUrl.js
+pause
